machine: Drop UBOOT_RD_ADDRESS from DHCOM PDK2
The UBOOT_RD_ADDRESS is optional and is inserted into fitImage loadaddr
field of the ramdisk image node. It is not necessary to specify this
here, as the ramdisk in the fitImage is relocated to the end of DRAM
from whichever location the fitImage was loaded to. The important part
is to make sure the ramdisk destination location does not interfere
with DTO application, by setting initrd_high below UBOOT_DTB_LOADADDRESS,
which is now done in the meta-dhsom-stm32-bsp boot.scr script.

Signed-off-by: Marek Vasut <marex@denx.de>