machine: Deduplicate the -rt machines
By including the non-rt machine configurations and overriding only the
relevant parts in conf/machine/include/dh-stm32mp1-common-rt.inc once,
which in this case are U-Boot and Linux version and kernel config name,
it is possible to significantly reduce the content of the -rt machines.

However, the decision whether or not the kernel should be patched by
preempt-rt patch should be left to the user and implemented entirely
in user local.conf. There is no reason to have separate machine config
for that, as machine config defines the physical machine and that does
not change for rt and non-rt patched kernel.

Signed-off-by: Marek Vasut <marex@denx.de>